requirements.txt
apache-airflow-providers-amazon
apache-airflow-providers-http
psycopg2-binary
requests
SQLAlchemy
pandas
openpyxl